% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bumper.R
\name{bumper}
\alias{bumper}
\title{bumper}
\usage{
bumper(GUID, variant = NULL)
}
\arguments{
\item{GUID}{character string the GUID of the content to trigger. You can find this value in the URL of your content, it should looks something like this: "b824db78-07b8-4205-b29e-0dbea32b4d8a"}

\item{variant}{optional numeric value of parameterized content to render. These are the numbers following the last / in the url of the content variant in Connect. If no value is specified the default variant will be rendered (for non-parameterized content the default is the only variant so there would be no need to use this argument).}
}
\value{
returns message with details if error or confirmation of success
}
\description{
trigger rendering of markdown jobs/reports and their variants hosted in Connect

The bumper function is intended to be called when you want to trigger an existing Connect job or report to run as a result of another process.

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge, instructions can be found here: \url{https://rstudioconnect.analytics.kellogg.com/RStudio_Knowledge_Library/}
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
bumper(GUID = "6dd37913-30fd-4c89-9c1d-4afd9dbfc6fe", variant = 595)
bumper(GUID = "6dd37913-30fd-4c89-9c1d-4afd9dbfc6fe")
}
