% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_checker.R
\name{env_checker}
\alias{env_checker}
\title{env_checker}
\usage{
env_checker()
}
\value{
If current env matches known/defined environments, a string containing the name of that environment is returned
}
\description{
Function to check user's current environment
}
\examples{
env_checker()
}
