% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_hadoop.R
\name{klink_hadoop}
\alias{klink_hadoop}
\title{Kellogg Hadoop connections}
\usage{
klink_hadoop(environment, schema, connection_pane = TRUE)
}
\arguments{
\item{environment}{character string containing desired environment. Valid arguments are "PROD" or "DEV".}

\item{schema}{character string containing name of desired database. Any database in the specified Hadoop environment should be valid.}

\item{connection_pane}{logical indicating whether information about your connection should be added to the Connections pane, default is TRUE.}
}
\value{
Returns Hadoop DBI connection object
}
\description{
Enables users to create connections to Kellogg Hadoop databases with a simple function
Note: Currently, you can only connect to Hadoop DEV from the UAT Workbench/Connect servers and PROD from the PROD Workbench/Connect servers.

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
# Typical connection w/ connection pane
prod_example <- klink_hadoop(environment = "PROD", schema = "KNA_BW")

# Typical connection w/o connection pane
keystone_example <- klink_hadoop("DEV", "KNA_BW", connection_pane = FALSE)
}
