% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_postgres.R
\name{klink_postgres}
\alias{klink_postgres}
\title{Kellogg RDS PostgreSQL connections}
\usage{
klink_postgres(environment, database, server = NULL, connection_pane = TRUE)
}
\arguments{
\item{environment}{character string containing desired environment. Currently only "DEV" is supported.}

\item{database}{character string containing name of desired database. Any database in the specified postgres environment should be valid so long as the service account being used has adequate permissions.}

\item{server}{OPTIONAL character string containing exact server name in R format (beware of /s). If server argument is provided the function will attempt to pass specified value into the connection call.}

\item{connection_pane}{OPTIONAL logical indicating whether information about your connection should be added to the Connections pane, default is TRUE.}
}
\value{
Returns PostgreSQL DBI connection object
}
\description{
Enables users to create connections to Kellogg PostgreSQL databases

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
# With connection pane
dev_con <- klink_postgres(environment = "DEV", database = "postgres")

# Without connection pane
dev_con <- klink_postgres("DEV", "postgres", connection_pane = FALSE)
}
