% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_redshift.R
\name{klink_redshift}
\alias{klink_redshift}
\title{Kellogg redshift connections}
\usage{
klink_redshift(environment, database = NULL, server = NULL, connection_pane = TRUE)
}
\arguments{
\item{environment}{character string containing desired environment. Valid arguments are "PROD", "DEV", or "QA"}

\item{database}{character string containing name of desired database. As of now, there is only one known databse for each environment. As such, this argument is optional and currently unused.}

\item{server}{character string containing exact server name in R format (beware of /s). As of now, there is only one known server for each environment. As such, this argument is optional and currently unused.}

\item{connection_pane}{logical indicating whether information about your connection should be added to the Connections pane, default is TRUE.}
}
\value{
Returns redshift DBI connection object
}
\description{
Enables users to create connections to Kellogg redshift databases

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
# Connect w/ connection pane
red_dev <- klink_redshift(environment = "DEV")

# Connect w/o connection pane
red_dev <- klink_redshift(environment = "DEV", connection_pane = FALSE)

# Example queries
DBI::dbGetQuery(red_dev, "SELECT DISTINCT tablename FROM PG_TABLE_DEF")
DBI::dbGetQuery(red_dev, "SELECT TOP 10 * FROM fin_acctg_ops.fisc_cal_wk")
}
