% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_s3R.R
\name{klink_s3R}
\alias{klink_s3R}
\title{Kellogg s3 (foreign) connection}
\usage{
klink_s3R()
}
\value{
populates user environment with required role and settings to access the corresponding S3 environment
}
\description{
Defines system environment variables required to utilize aws.s3 functions and returns proper value for s3BucketName

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
# Retrieve required system settings (in background) and appropriate s3 bucket name
klink_s3R()

# Use example aws.s3 functions to retrieve information from s3 bucket
aws.s3::object_exists("/data/R_training/package_list.rds", bucket = s3BucketName)
aws.s3::object_size("/data/R_training/package_list.rds", bucket = s3BucketName)
}
