% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_scrub.R
\name{klink_scrub}
\alias{klink_scrub}
\title{klink_scrub}
\usage{
klink_scrub(data_to_scrub)
}
\arguments{
\item{data_to_scrub}{dataframe to be scrubbed (no quotes)}
}
\value{
Returns original data with NaN and Inf values replaced with NAs (which write to databases as NULL values)
}
\description{
removes NaN and Inf values from numeric and character fields, meant to make prevent associated errors in database write operations
}
\examples{
klink_scrub(my_df)
klink_scrub(another_df)
}
