% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_sql.R
\name{klink_sql}
\alias{klink_sql}
\title{Kellogg SQL connections}
\usage{
klink_sql(environment, database, server = NULL, connection_pane = TRUE)
}
\arguments{
\item{environment}{character string containing desired environment. Valid arguments are "PROD" or "DEV".}

\item{database}{optional character string containing name of desired database. Any database in the specified SQL environment should be valid.}

\item{server}{character string containing exact server name in R format (beware of /s). If no argument provided will attempt to use Keystone servers. If server argument is provided the function will attempt to pass specified value into the connection call.}

\item{connection_pane}{logical indicating whether information about your connection should be added to the Connections pane, default is TRUE.}
}
\value{
Returns SQL DBI connection object
}
\description{
Enables users to create connections to Kellogg SQL databases with a simple function

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
# Typical connection w/ connection pane
keystone_example <- klink_sql(environment = "DEV", database = "KG_ANALYTICS_APPS")
# Typical connection w/o connection pane
keystone_example <- klink_sql("DEV", "KG_ANALYTICS_APPS", connection_pane = FALSE)
# Connection to non-standard SQL server
# (these are supported on a case-by-case basis, contact data science team for support)
other__ex <- klink_sql("DEV", "KG_ANALYTICS_APPS", server = "USAWSCWSQL5066\\\\ANALYTICSDEV3")
}
