% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\alias{runKLINK}
\title{Launch KLINK}
\usage{
launchApp()

runKLINK()
}
\value{
No return value, called for side effects.
}
\description{
This launches the KLINK app. \code{runKLINK()} is a synonym for \code{launchApp()}.
}
\examples{

\dontrun{
launchApp()
}

}
