% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LINKAGEMAP}
\alias{LINKAGEMAP}
\title{Built-in linkage map}
\format{
A data frame with 50 rows and 4 columns: \code{Marker}, \code{Kit}, \code{Chrom},
\code{PosCM}.
}
\usage{
LINKAGEMAP
}
\description{
A genetic map including 50 autosomal STR markers.
}
\keyword{datasets}
