% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLINK-package.R
\docType{data}
\name{map50}
\alias{map50}
\title{Genetic marker map}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 4 columns.
}
\usage{
map50
}
\description{
This object is re-exported from \link[norSTR:map50]{norSTR::map50}.
}
\keyword{datasets}
