% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerSummary.R
\name{markerSummary}
\alias{markerSummary}
\title{Generate table of marker data}
\usage{
markerSummary(pedigrees, replaceNames = FALSE)
}
\arguments{
\item{pedigrees}{A list of 2 pedigrees.}

\item{replaceNames}{A logical, indicating if IDs should be changed to
Person1, Person2, ...}
}
\value{
A data frame.
}
\description{
Generate table of marker data
}
\examples{
markerSummary(paternity)

}
