% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_exists.R
\name{object_exists}
\alias{object_exists}
\title{object_exists}
\usage{
object_exists(object_name)
}
\arguments{
\item{object_name}{name of object (without quotation marks)}
}
\value{
If object exists will return TRUE else FALSE
}
\description{
checks for existence of object when name passed without quotes
}
\examples{
object_exists(test)
object_exists(test2)
}
