% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{parseXML}
\alias{parseXML}
\title{Parse XML file associated with .fam file}
\usage{
parseXML(xml, famname = NULL, famids = NULL)
}
\arguments{
\item{xml}{Path to a file with extension .xml.}

\item{famname}{(Optional) The basename of the associated .fam file, used to
test that the files belong together.}

\item{famids}{(Optional) A character of the ID strings recorded in the .fam
file. These should also be present in the XML file.}
}
\description{
Parse XML file associated with .fam file
}
\examples{
# (No example included)

}
