% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paternity}
\alias{paternity}
\title{Dataset for a paternity case}
\format{
A list of two pedigrees, named H1 and H2.
}
\usage{
paternity
}
\description{
A list of two pedigrees forming the hypotheses in a paternity case: H1 (AF is
the father of CH) and H2 (unrelated). AF and CH are genotyped with 11 marker.
}
\examples{
library(forrel)

plotPedList(paternity, marker = "SE33")
kinshipLR(paternity)
}
\keyword{datasets}
