% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_object_exists.R
\name{r_object_exists}
\alias{r_object_exists}
\title{r_object_exists}
\usage{
r_object_exists(object_name)
}
\arguments{
\item{object_name}{name of object (without quotation marks)}
}
\value{
If object exists will return TRUE else FALSE
}
\description{
checks for existence of object when name passed without quotes
}
\examples{
r_object_exists(test)
r_object_exists(test2)
}
