% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKLINK.R
\name{runKLINK}
\alias{runKLINK}
\title{Launch KLINK}
\usage{
runKLINK()
}
\value{
No return value, called for side effects.
}
\description{
This launches the shiny app KLINK.
}
\examples{

\dontrun{
runKLINK()
}

}
