% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoltar.R
\name{zoltar}
\alias{zoltar}
\title{zoltar}
\usage{
zoltar(token, solution_name = NULL)
}
\arguments{
\item{token}{character string containing the exact name of secret to be retrieved (if passing value directly must be wrapped in quotation marks to be interpreted as character string)}

\item{solution_name}{character string identifying the name of the solution where this logic is being used, for recording to logs}
}
\value{
If name is found, a string containing the secret value else a string stating that zoltar is unable to grant your wish
}
\description{
Allows internal users with Connect API keys to reference zoltar API and retrieve secrets

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge, instructions can be found here: \url{https://rstudioconnect.analytics.kellogg.com/RStudio_Knowledge_Library/}
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
zoltar("s3BucketName")
zoltar("MS_SQL_ANALYTICS_DEV_server")
}
