% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoltar_list.R
\name{zoltar_list}
\alias{zoltar_list}
\title{Zoltar variables list}
\usage{
zoltar_list()
}
\value{
Returns list of valid zoltar API requests
}
\description{
Enables users to retrieve list of all valid zoltar API requests

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge).
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
zoltar_list()
}
