% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuul.R
\name{zuul}
\alias{zuul}
\title{zuul}
\usage{
zull()
}
\value{
data frame containing usage details
}
\description{
Logs zoltar usage

In order to use these tools users must first:
\enumerate{
\item Have a RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge, instructions can be found here: \url{https://rstudioconnect.analytics.kellogg.com/RStudio_Knowledge_Library/}
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}
\examples{
zuul()
}
