% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLINK-package.R
\docType{package}
\name{KLINK-package}
\alias{KLINK}
\alias{KLINK-package}
\title{KLINK: Kinship Analysis with Linked Markers}
\description{
A Shiny app for computing kinship LR using pairwise linked markers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/KLINK}
  \item Report bugs at \url{https://github.com/magnusdv/KLINK/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no}

}
\keyword{internal}
