% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLINK-package.R
\docType{package}
\name{KLINK-package}
\alias{KLINK}
\alias{KLINK-package}
\title{KLINK: Kinship Analysis With Linked Markers}
\description{
A Shiny app for computing kinship LR using linked markers.
}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{magnusdv@medisin.uio.no}

}
\keyword{internal}
