% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLINK-package.R
\docType{package}
\name{KLINK-package}
\alias{KLINK}
\alias{KLINK-package}
\title{KLINK: Kinship Analysis with Linked Markers}
\description{
A 'shiny' application for forensic kinship testing, based on the 'pedsuite' R packages. 'KLINK' is closely aligned with the (non-R) software 'Familias' and 'FamLink', but offers several unique features, including visualisations and automated report generation. The calculation of likelihood ratios supports pairs of linked markers, and all common mutation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/KLINK}
  \item Report bugs at \url{https://github.com/magnusdv/KLINK/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
