% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLINK-package.R
\docType{package}
\name{KLINK-package}
\alias{KLINK}
\alias{KLINK-package}
\title{KLINK: Kinship Analysis with Linked Markers}
\description{
A Shiny app based on the 'pedsuite' packages, for computing likelihood ratios in forensic kinship testing. The app is similar in purpose to the (non-R) software 'Familias' and 'FamLink', but offers a number of unique features, including visualisations and automated report generation. Pairs of linked markers are supported, as well as all common mutation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/KLINK}
  \item Report bugs at \url{https://github.com/magnusdv/KLINK/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no}

}
\keyword{internal}
