% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink.R
\docType{package}
\name{klink}
\alias{klink}
\title{klink is an internal R package intended to provide common helper functions that may be of value to Kellogg users.}
\description{
Most of the current functions are focused on easily forming connections to internal data connections by removing the necessity of users to maintain their own local secrets for commonly used platforms and service accounts.
}
\details{
This is accomplished through the use of a centrally managed API hosted in our internal RStudio Connect environment.
}
\section{klink functions}{

The klink functions are a general group of functions used to establish links to data sources commonly used within Kellogg.

Currently these functions include:
\itemize{
\item klink_sql (for SQL connections)
\item klink_s3R (for our foreign s3 bucket connections)
}

See vignette or function documentaion (?klink_sql or ?klink_s3R) for more information and examples.
}

\section{zoltar function}{

The klink functions are esssentially wrappers that utilize zoltar under the hood to make the user experience as lightweight as possible.

When a wish/alias of a known value is passed to this function our internal zoltar API returns the requested value. The Connect API key required in the user setup allows us to secure this functionality while also centrally managing our shared service account credentials. In addition to reduced upkeep for users this also limits the exposure to the underlying secrets associated with these account.

If there is a service account that isn't for SQL, s3, or similar you may want to try to use zoltar directly. If there are service accounts that are not yet supported by zoltar please reach out to the Kellogg Data Science team to have them added.

See vignette or function documentation (?zoltar) for more information and examples.
}

\section{Setup}{

In order to use these functions users must first:
\enumerate{
\item Have an RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge, instructions can be found here: \url{https://rstudioconnect.analytics.kellogg.com/RStudio_Knowledge_Library/}
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
}
}

