% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klink_sql.R
\name{klink_sql}
\alias{klink_sql}
\title{Kellogg SQL connections}
\usage{
klink_sql(environment, database)
}
\arguments{
\item{environment}{Valid arguments are "PROD" or "DEV"}

\item{database}{Any database in the specified SQL environment should be valid. In current version values of  KG_R_APPS will attempt to use KG_R_APPS svc account credentials, KG_SAS will use KG_SAS credentials, and all other databases will attempt to use KG_ANALYTICS_APPS. If you have suggestions for refining or correcting these functions please contact package authors.}
}
\value{
Returns SQL odbc connection object
}
\description{
Enables internal users with Connect API keys to create connections to Kellogg SQL databases with a simple function

In order to use these tools users must first:
\enumerate{
\item Have an RStudio Connect account (you likely have one already if you're using RStudio Workbench, if not you can request access through Digital Concierge, instructions can be found here: \url{https://rstudioconnect.analytics.kellogg.com/RStudio_Knowledge_Library/}
\item Create a local RStudio Connect API key \url{https://docs.rstudio.com/connect/user/api-keys}
\item Create an .Renviron file in your Home folder assigning your API key value to the name CONNECT_API_KEY \url{https://rstats.wtf/r-startup.html}
\item Install DBI and odbc packages
}
}
\examples{
conn <- klink_sql("DEV", "KG_ANALYTICS_APPS")
}
