% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LINKAGEMAP}
\alias{LINKAGEMAP}
\title{Built-in linkage map}
\format{
A data frame with 18 rows and 5 columns.
}
\usage{
LINKAGEMAP
}
\description{
A genetic map including 9 pairs of linked STR markers
}
\keyword{datasets}
