% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerSummary.R
\name{markerSummary}
\alias{markerSummary}
\title{Generate table of marker data}
\usage{
markerSummary(pedigrees, linkageMap = NULL)
}
\arguments{
\item{pedigrees}{A list of 2 pedigrees.}

\item{linkageMap}{A data frame.}
}
\value{
A data frame.
}
\description{
Generate table of marker data
}
\examples{
markerSummary(paternity)

}
