% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{parseXML}
\alias{parseXML}
\title{Parse \code{XML} file associated with \code{.fam} file}
\usage{
parseXML(xml)
}
\arguments{
\item{xml}{Path to a file with extension \code{.xml}.}
}
\description{
Parse \code{XML} file associated with \code{.fam} file
}
\examples{
# (No example included)

}
