% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{parseXML}
\alias{parseXML}
\title{Parse XML file associated with .fam file}
\usage{
parseXML(xml)
}
\arguments{
\item{xml}{Path to a file with extension .xml.}
}
\description{
Parse XML file associated with .fam file
}
\examples{
# (No example included)

}
