% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paternity}
\alias{paternity}
\title{Dataset for a paternity case}
\format{
A list of two pedigrees, named H1 and H2.
}
\usage{
paternity
}
\description{
A list of two pedigrees forming the hypotheses in a paternity case: H1 (\code{AF} is
the father of \code{CH}) and H2 (unrelated). \code{AF} and \code{CH} are genotyped with 11
markers, with allele frequencies from \code{forrel::NorwegianFrequencies}.
}
\examples{
pedtools::plotPedList(paternity, marker = "SE33")
forrel::kinshipLR(paternity)
}
\keyword{datasets}
